# [dipu][optim] speedup mm by register dipu_mm and use nodispatch::empty