--- a/dipu/scripts/autogen_diopi_wrapper/diopi_wrapper_template.py
+++ b/dipu/scripts/autogen_diopi_wrapper/diopi_wrapper_template.py
@@ -1,19 +1,58 @@
 # Copyright (c) 2023, DeepLink.
 diopi_wrapper_file_template_content = \
 """// autogened file
+#include <algorithm>
+#include <array>
+#include <cmath>
+#include <cstddef>
+#include <cstdint>
+#include <cstdio>
+#include <iostream>
+#include <ostream>
+#include <tuple>
+#include <utility>
 #include <vector>
 
-#include <ATen/ATen.h>
+#include <ATen/ExpandUtils.h>
 #include <ATen/Functions.h>
 #include <ATen/Tensor.h>
+#include <ATen/core/ATen_fwd.h>
+#include <ATen/core/Generator.h>
+#include <ATen/core/LegacyTypeDispatch.h>
+#include <ATen/core/List.h>
+#include <ATen/core/TensorBody.h>
 #include <ATen/native/ReduceOpsUtils.h>
+#include <ATen/ops/empty.h>
+#include <ATen/ops/empty_like.h>
+#include <ATen/ops/scalar_tensor.h>
+#include <ATen/ops/to_native.h>
+#include <ATen/ops/zeros.h>
+#include <c10/core/MemoryFormat.h>
+#include <c10/core/ScalarType.h>
+#include <c10/core/SymInt.h>
+#include <c10/core/SymIntArrayRef.h>
+#include <c10/util/ArrayRef.h>
+#include <c10/util/Exception.h>
+#include <c10/util/Optional.h>
+#include <c10/util/SmallVector.h>
+#include <c10/util/accumulate.h>
+#include <c10/util/string_view.h>
 #include <torch/csrc/autograd/custom_function.h>
+#include <torch/csrc/autograd/generated/variable_factories.h>
 #include <torch/types.h>
+
+#include <diopi/diopirt.h>
+#include <diopi/functions.h>
 
 #include "csrc_dipu/aten/RegisterDIPU.hpp"
 #include "csrc_dipu/aten/ops/DIPUCopy.hpp"
+#include "csrc_dipu/aten/ops/NodispatchUtils.hpp"
+#include "csrc_dipu/aten/ops/OpUtils.hpp"
+#include "csrc_dipu/base/basedef.h"
 #include "csrc_dipu/diopirt/diopirt_impl.h"
 #include "csrc_dipu/profiler/profiler.h"
+#include "csrc_dipu/runtime/core/DIPUGeneratorImpl.h"
+#include "csrc_dipu/runtime/core/DIPUStream.h"
 
 #include "CustomFallbackFunctions.hpp"