--- a/dipu/tests/python/unittests/test_mm.py
+++ b/dipu/tests/python/unittests/test_mm.py
@@ -9,7 +9,7 @@ class TestMm(TestCase):
         dipu = torch.device("dipu")
         cpu = torch.device("cpu")
         mat1 = torch.randn(2, 3)
-        mat2 = torch.randn(3, 3)
+        mat2 = torch.randn(3, 4)
         r1 = torch.mm(mat1.to(dipu), mat2.to(dipu))
         r2 = torch.mm(mat1.to(cpu), mat2.to(cpu))
         self.assertEqual(r1.to(cpu), r2)