--- a/dipu/torch_dipu/csrc_dipu/aten/DIPUATenFunctions.h
+++ b/dipu/torch_dipu/csrc_dipu/aten/DIPUATenFunctions.h
@@ -1,62 +1,65 @@
 // Copyright (c) 2023, DeepLink.
 #pragma once
 
-#include <ATen/ATen.h>
-#include <ATen/Dispatch.h>
-#include <ATen/Tensor.h>
+#include <cstddef>
+#include <cstdint>
+
+#include <ATen/core/ATen_fwd.h>
+#include <ATen/core/TensorBody.h>
+#include <c10/core/Device.h>
+#include <c10/core/Layout.h>
+#include <c10/core/MemoryFormat.h>
+#include <c10/core/ScalarType.h>
+#include <c10/core/Storage.h>
+#include <c10/core/StorageImpl.h>
+#include <c10/util/Optional.h>
 
 namespace dipu {
 namespace native {
+namespace dipu_aten {
+// dipu native func
+at::Tensor empty(at::IntArrayRef size, c10::optional<at::ScalarType> dtype_opt,
+                 c10::optional<at::Layout> layout_opt,
+                 c10::optional<at::Device> device_opt,
+                 c10::optional<bool> pin_memory_opt,
+                 c10::optional<at::MemoryFormat> memory_format_opt);
+at::Tensor empty_cpu(at::IntArrayRef size,
+                     c10::optional<at::ScalarType> dtype_opt,
+                     c10::optional<at::Layout> layout_opt,
+                     c10::optional<at::Device> device_opt,
+                     c10::optional<bool> pin_memory_opt,
+                     c10::optional<at::MemoryFormat> memory_format_opt);
 
-struct DIPUATenFunctions {
-  // dipu native func
-  static at::Tensor empty(at::IntArrayRef size,
-                          c10::optional<at::ScalarType> dtype_opt,
-                          c10::optional<at::Layout> layout_opt,
-                          c10::optional<at::Device> device_opt,
-                          c10::optional<bool> pin_memory_opt,
-                          c10::optional<at::MemoryFormat> memory_format_opt);
-  static at::Tensor empty_cpu(
-      at::IntArrayRef size, c10::optional<at::ScalarType> dtype_opt,
-      c10::optional<at::Layout> layout_opt,
-      c10::optional<at::Device> device_opt, c10::optional<bool> pin_memory_opt,
-      c10::optional<at::MemoryFormat> memory_format_opt);
+at::Tensor empty_strided(at::IntArrayRef size, at::IntArrayRef stride,
+                         c10::optional<at::ScalarType> dtype_opt,
+                         c10::optional<at::Layout> layout_opt,
+                         c10::optional<at::Device> device_opt,
+                         c10::optional<bool> pin_memory_opt);
+at::Tensor empty_strided_cpu(at::IntArrayRef size, at::IntArrayRef stride,
+                             c10::optional<at::ScalarType> dtype_opt,
+                             c10::optional<at::Layout> layout_opt,
+                             c10::optional<at::Device> device_opt,
+                             c10::optional<bool> pin_memory_opt);
 
-  static at::Tensor empty_strided(at::IntArrayRef size, at::IntArrayRef stride,
-                                  c10::optional<at::ScalarType> dtype_opt,
-                                  c10::optional<at::Layout> layout_opt,
-                                  c10::optional<at::Device> device_opt,
-                                  c10::optional<bool> pin_memory_opt);
-  static at::Tensor empty_strided_cpu(at::IntArrayRef size,
-                                      at::IntArrayRef stride,
-                                      c10::optional<at::ScalarType> dtype_opt,
-                                      c10::optional<at::Layout> layout_opt,
-                                      c10::optional<at::Device> device_opt,
-                                      c10::optional<bool> pin_memory_opt);
+const at::Tensor& resize_(const at::Tensor& self, at::IntArrayRef size,
+                          c10::optional<at::MemoryFormat> memory_format);
 
-  static const at::Tensor& resize_(
-      const at::Tensor& self, at::IntArrayRef size,
-      c10::optional<at::MemoryFormat> memory_format);
+at::Scalar _local_scalar_dense_dipu(const at::Tensor& self);
 
-  static at::Scalar _local_scalar_dense_dipu(const at::Tensor& self);
+at::Tensor& set_storage_dipu_(at::Tensor& result, c10::Storage storage,
+                              int64_t storage_offset, at::IntArrayRef size,
+                              at::IntArrayRef stride);
+at::Tensor& set_dipu_(at::Tensor& self);
 
-  static at::Tensor& set_storage_dipu_(at::Tensor& result, c10::Storage storage,
-                                       int64_t storage_offset,
-                                       at::IntArrayRef size,
-                                       at::IntArrayRef stride);
-  static at::Tensor& set_dipu_(at::Tensor& self);
+void resize_bytes_dipu(c10::StorageImpl* storage, size_t newsize_bytes);
 
-  static void resize_bytes_dipu(c10::StorageImpl* storage,
-                                size_t newsize_bytes);
+bool is_pinned(const at::Tensor& self, c10::optional<at::Device> device);
+at::Tensor _pin_memory(const at::Tensor& self,
+                       c10::optional<at::Device> device);
 
-  static bool is_pinned(const at::Tensor& self,
-                        c10::optional<at::Device> device);
-  static at::Tensor _pin_memory(const at::Tensor& self,
-                                c10::optional<at::Device> device);
-
-  // todo:: use same format as autogen
-  // diopi function defined in AutoGenedKernels.cpp,
-};
+// todo:: use same format as autogen
+// diopi function defined in AutoGenedKernels.cpp,
+};  // namespace dipu_aten
 
 }  // namespace native
 }  // namespace dipu