--- a/dipu/torch_dipu/csrc_dipu/aten/RegisterDIPU.cpp
+++ b/dipu/torch_dipu/csrc_dipu/aten/RegisterDIPU.cpp
@@ -14,7 +14,7 @@
 #include <csrc_dipu/base/basedef.h>
 #include <csrc_dipu/profiler/profiler.h>
 
-using dnative = dipu::native::DIPUATenFunctions;
+namespace dnative = dipu::native::dipu_aten;
 
 namespace dipu {
 namespace {