--- a/dipu/torch_dipu/csrc_dipu/aten/ops/EmptyOpsKernel.cpp
+++ b/dipu/torch_dipu/csrc_dipu/aten/ops/EmptyOpsKernel.cpp
@@ -1,13 +1,25 @@
 // Copyright (c) 2023, DeepLink.
 #include <ATen/EmptyTensor.h>
+#include <ATen/core/ATen_fwd.h>
+#include <c10/core/Allocator.h>
 #include <c10/core/CPUAllocator.h>
-#include <c10/core/TensorImpl.h>
+#include <c10/core/Device.h>
+#include <c10/core/DeviceType.h>
+#include <c10/core/DispatchKey.h>
+#include <c10/core/DispatchKeySet.h>
+#include <c10/core/Layout.h>
+#include <c10/core/MemoryFormat.h>
+#include <c10/core/ScalarType.h>
 #include <c10/core/TensorOptions.h>
+#include <c10/util/Exception.h>
+#include <c10/util/Optional.h>
 #include <c10/util/accumulate.h>
 
-#include <csrc_dipu/aten/DIPUATenFunctions.h>
-#include <csrc_dipu/profiler/profiler.h>
-#include <csrc_dipu/runtime/rthelper.h>
+#include "csrc_dipu/aten/DIPUATenFunctions.h"
+#include "csrc_dipu/base/basedef.h"
+#include "csrc_dipu/profiler/profiler.h"
+#include "csrc_dipu/runtime/core/allocator/DIPUCachingAllocator.h"
+#include "csrc_dipu/runtime/rthelper.h"
 
 namespace dipu {
 namespace native {
@@ -19,11 +31,12 @@ static c10::Allocator* GetCPUAllocatorMaybePinned(bool pin_memory) {
   return c10::GetCPUAllocator();
 }
 
-at::Tensor DIPUATenFunctions::empty(
-    at::IntArrayRef size, c10::optional<at::ScalarType> dtype_opt,
-    c10::optional<at::Layout> layout_opt, c10::optional<at::Device> device_opt,
-    c10::optional<bool> pin_memory_opt,
-    c10::optional<at::MemoryFormat> memory_format_opt) {
+at::Tensor dipu_aten::empty(at::IntArrayRef size,
+                            c10::optional<at::ScalarType> dtype_opt,
+                            c10::optional<at::Layout> layout_opt,
+                            c10::optional<at::Device> device_opt,
+                            c10::optional<bool> pin_memory_opt,
+                            c10::optional<at::MemoryFormat> memory_format_opt) {
   dipu::profile::RecordBlockCreator dipu_recorder(__FUNCTION__);
   TORCH_INTERNAL_ASSERT_DEBUG_ONLY(c10::device_or_default(device_opt).type() ==
                                    dipu::DIPU_DEVICE_TYPE);
@@ -37,7 +50,7 @@ at::Tensor DIPUATenFunctions::empty(
                                    memory_format_opt);
 }
 
-at::Tensor DIPUATenFunctions::empty_cpu(
+at::Tensor dipu_aten::empty_cpu(
     at::IntArrayRef size, c10::optional<at::ScalarType> dtype_opt,
     c10::optional<at::Layout> layout_opt, c10::optional<at::Device> device_opt,
     c10::optional<bool> pin_memory_opt,
@@ -56,11 +69,12 @@ at::Tensor DIPUATenFunctions::empty_cpu(
 }
 
 // use empty_generic, test
-at::Tensor DIPUATenFunctions::empty_strided(
-    at::IntArrayRef size, at::IntArrayRef stride,
-    c10::optional<at::ScalarType> dtype_opt,
-    c10::optional<at::Layout> layout_opt, c10::optional<at::Device> device_opt,
-    c10::optional<bool> pin_memory_opt) {
+at::Tensor dipu_aten::empty_strided(at::IntArrayRef size,
+                                    at::IntArrayRef stride,
+                                    c10::optional<at::ScalarType> dtype_opt,
+                                    c10::optional<at::Layout> layout_opt,
+                                    c10::optional<at::Device> device_opt,
+                                    c10::optional<bool> pin_memory_opt) {
   dipu::profile::RecordBlockCreator dipu_recorder(__FUNCTION__);
   auto device = c10::device_or_default(device_opt);
   AT_ASSERT(device.type() == dipu::DIPU_DEVICE_TYPE);
@@ -73,11 +87,12 @@ at::Tensor DIPUATenFunctions::empty_strided(
                                            dtype);
 }
 
-at::Tensor DIPUATenFunctions::empty_strided_cpu(
-    at::IntArrayRef size, at::IntArrayRef stride,
-    c10::optional<at::ScalarType> dtype_opt,
-    c10::optional<at::Layout> layout_opt, c10::optional<at::Device> device_opt,
-    c10::optional<bool> pin_memory_opt) {
+at::Tensor dipu_aten::empty_strided_cpu(at::IntArrayRef size,
+                                        at::IntArrayRef stride,
+                                        c10::optional<at::ScalarType> dtype_opt,
+                                        c10::optional<at::Layout> layout_opt,
+                                        c10::optional<at::Device> device_opt,
+                                        c10::optional<bool> pin_memory_opt) {
   TORCH_INTERNAL_ASSERT_DEBUG_ONLY(c10::device_or_default(device_opt).type() ==
                                    c10::DeviceType::CPU);
   TORCH_INTERNAL_ASSERT_DEBUG_ONLY(c10::layout_or_default(layout_opt) ==