--- a/dipu/torch_dipu/csrc_dipu/aten/ops/OpUtils.hpp
+++ b/dipu/torch_dipu/csrc_dipu/aten/ops/OpUtils.hpp
@@ -1,5 +1,25 @@
 #pragma once
 
+#include <array>
+#include <cstddef>
+#include <cstdint>
+#include <cstdlib>
+#include <sstream>
+#include <string>
+#include <vector>
+
+#include <ATen/core/ATen_fwd.h>
+#include <ATen/core/Generator.h>
+#include <ATen/core/List.h>
+#include <ATen/core/TensorBody.h>
+#include <ATen/ops/abs.h>
+#include <ATen/ops/allclose.h>
+#include <c10/util/ArrayRef.h>
+#include <c10/util/Optional.h>
+#include <c10/util/OptionalArrayRef.h>
+#include <c10/util/string_view.h>
+
+#include "csrc_dipu/runtime/core/DIPUStream.h"
 #include <csrc_dipu/runtime/rthelper.h>
 #include <csrc_dipu/utils/Log.h>
 
@@ -7,7 +27,7 @@ namespace dipu {
 namespace native {
 
 inline bool checkTensorDevice() {
-  static bool enable = []() {
+  bool enable = []() {
     const char* env_ptr = std::getenv("DIPU_CHECK_TENSOR_DEVICE");
     if (env_ptr == nullptr) {
       return false;
@@ -18,31 +38,31 @@ inline bool checkTensorDevice() {
 }
 
 inline void synchronizeIfEnable() {
-  static const char* mode = std::getenv("DIPU_SYNC_EXEC_MODE");
+  const char* mode = std::getenv("DIPU_SYNC_EXEC_MODE");
   if (mode != nullptr) {
     DIPU_LOG_ONCE << "The synchronous operation is performed after "
                   << "the diopi function call because the DIPU_SYNC_EXEC_MODE "
                      "environment variable is set"
-                  << std::endl;
+                  << '\n';
     dipu::getCurrentDIPUStream().synchronize();
   }
 }
 
 inline int dumpOpArgLevel() {
-  static const char* env_ptr = std::getenv("DIPU_DUMP_OP_ARGS");
-  static int level = env_ptr ? std::atoi(env_ptr) : 0;
+  const char* env_ptr = std::getenv("DIPU_DUMP_OP_ARGS");
+  int level = env_ptr ? std::atoi(env_ptr) : 0;
   return level;
 }
 
 template <typename T>
-static std::string dumpArg(const T& t) {
+std::string dumpArg(const T& t) {
   std::stringstream stream;
   stream << t;
   return stream.str();
 }
 
 template <typename T1>
-static std::string dumpArg(const c10::optional<T1>& opt_t) {
+std::string dumpArg(const c10::optional<T1>& opt_t) {
   std::stringstream stream;
   if (opt_t.has_value()) {
     stream << dumpArg(opt_t.value());
@@ -51,7 +71,7 @@ static std::string dumpArg(const c10::optional<T1>& opt_t) {
 }
 
 template <typename T>
-static std::string dumpArg(const c10::OptionalArrayRef<T>& opt_t) {
+std::string dumpArg(const c10::OptionalArrayRef<T>& opt_t) {
   std::stringstream stream;
   if (opt_t.has_value()) {
     stream << dumpArg(opt_t.value());
@@ -60,7 +80,7 @@ static std::string dumpArg(const c10::OptionalArrayRef<T>& opt_t) {
 }
 
 template <typename T1, template <typename elem> class container>
-static std::string dumpArg(const container<T1>& t) {
+std::string dumpArg(const container<T1>& t) {
   std::stringstream stream;
   for (auto iter = t.begin(); iter != t.end(); ++iter) {
     stream << dumpArg(*iter) << ", ";
@@ -69,7 +89,7 @@ static std::string dumpArg(const container<T1>& t) {
 }
 
 template <>
-std::string dumpArg(const at::Tensor& tensor) {
+inline std::string dumpArg(const at::Tensor& tensor) {
   std::stringstream stream;
   if (tensor.defined()) {
     stream << "numel: " << tensor.numel() << ", sizes: " << tensor.sizes()
@@ -81,7 +101,7 @@ std::string dumpArg(const at::Tensor& tensor) {
            << ", memory_format: " << tensor.suggest_memory_format()
            << ",  data_ptr: " << tensor.data_ptr();
     if (dumpOpArgLevel() > 2) {
-      stream << std::endl << tensor;
+      stream << '\n' << tensor;
     }
   } else {
     stream << "undefined";
@@ -90,24 +110,24 @@ std::string dumpArg(const at::Tensor& tensor) {
 }
 
 template <>
-std::string dumpArg(const at::Scalar& t) {
+inline std::string dumpArg(const at::Scalar& t) {
   std::stringstream stream;
   stream << t;
   return stream.str();
 }
 
 template <>
-std::string dumpArg(const c10::string_view& t) {
+inline std::string dumpArg(const c10::string_view& t) {
   return dumpArg(std::string(t.data()));
 }
 
 template <>
-std::string dumpArg(const at::Generator& t) {
+inline std::string dumpArg(const at::Generator& t) {
   return "";
 }
 
 template <typename T, size_t N>
-static std::string dumpArg(const std::array<T, N>& t) {
+std::string dumpArg(const std::array<T, N>& t) {
   std::stringstream stream;
   for (auto iter = t.begin(); iter != t.end(); ++iter) {
     stream << dumpArg(*iter) << " ";
@@ -116,25 +136,25 @@ static std::string dumpArg(const std::array<T, N>& t) {
 }
 
 template <>
-std::string dumpArg(const c10::List<c10::optional<at::Tensor>>& t) {
+inline std::string dumpArg(const c10::List<c10::optional<at::Tensor>>& t) {
   std::stringstream stream;
-  stream << "size:" << t.size() << std::endl;
+  stream << "size:" << t.size() << '\n';
   for (int i = 0; i < t.size(); ++i) {
     bool has_value = t[i].has_value();
     stream << "\t" << i << "th: has_value:" << has_value << " ";
     if (has_value) {
       stream << dumpArg(t[i].value());
     }
-    stream << std::endl;
+    stream << '\n';
   }
   return stream.str();
 }
 
 template <typename T1, typename T2, template <typename elem1> class container1,
           template <typename elem2> class container2>
-static std::vector<int64_t> infer_reduce_op_shape(
-    const container1<T1>& input_shape, const container2<T2>& dims,
-    bool keepdim) {
+std::vector<int64_t> infer_reduce_op_shape(const container1<T1>& input_shape,
+                                           const container2<T2>& dims,
+                                           bool keepdim) {
   if (dims.size() <= 0) {
     return {};
   }
@@ -166,7 +186,7 @@ static std::vector<int64_t> infer_reduce_op_shape(
   return output_shape;
 }
 
-static std::string _allclose(const at::Tensor& a, const at::Tensor& b) {
+inline std::string _allclose(const at::Tensor& a, const at::Tensor& b) {
   if (a.defined() && b.defined()) {
     try {
       constexpr double tolerance_absolute = 1e-4;
@@ -191,7 +211,7 @@ static std::string _allclose(const at::Tensor& a, const at::Tensor& b) {
   }
 }
 
-static std::string _allclose(const c10::ArrayRef<at::Tensor>& a,
+inline std::string _allclose(const c10::ArrayRef<at::Tensor>& a,
                              const c10::ArrayRef<at::Tensor>& b) {
   if (a.size() != b.size()) {
     return "not_allclose:";