… dipu/scripts/autogen_diopi_wrapper/diopi_functions.yaml
… dipu/torch_dipu/csrc_dipu/vendor/cuda/patch/wrapperRegister.cpp