mm [4, 4096]x[4096, 4096]: 39us -> 38us -> 36.1us

llama7b eval 123ms->121ms

plus some refactor and include cleaning